Address PR review: fix resolution consumers and cleanup
- Replace _config.resolution with require_uniform_grid() in width_sweep_analysis.py
- Remove float() casts on traced JAX arrays in DiffractiveDetector
- Use jnp.float32 for face area weights in ModeOverlapDetector (real-valued geometry)
- Remove unused widths variable in PerfectlyMatchedLayer non-uniform depth calc

+++ src/fdtdx/objects/detectors/mode.py
@@ -102,7 +102,7 @@ class ModeOverlapDetector(PhasorDetector):
             return grid.face_area(axis=self.propagation_axis, slice_tuple=self.grid_slice_tuple)
 
         spacing = self._config.require_uniform_grid()
-        return jnp.ones(self.grid_shape, dtype=self.dtype) * spacing * spacing
+        return jnp.ones(self.grid_shape, dtype=jnp.float32) * spacing * spacing
 
     def _face_area_weights(self) -> jax.Array:
         """Return cached detector-plane face areas for mode-overlap integration."""
